deleted list controller (duplicate of item controller), updated abstract entity to remove 'name' field, fixed errors created within ItemData class.
