+++ src/main/java/org/launchcode/inventorytrackergradle/controllers/ItemController.java
@@ -2,17 +2,14 @@ package org.launchcode.inventorytrackergradle.controllers;
 
 import org.launchcode.inventorytrackergradle.models.Item;
 import org.launchcode.inventorytrackergradle.models.data.ItemRepository;
-import org.launchcode.inventorytrackergradle.models.data.ManufacturerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-
 @RestController
+@CrossOrigin(origins = "http://localhost:4200")
 @RequestMapping(value="items")
 public class ItemController {
     @Autowired
