+++ src/main/java/org/launchcode/inventorytrackergradle/controllers/UserController.java
@@ -14,18 +14,6 @@ public class UserController {
     @Autowired
     private UserRepository userRepository;
 
-    @GetMapping("populate")
-    String populateDatabase(){
-
-        User user = new User("NewUser1", "mail@email.com", "1234567", "1234", "1234");
-        userRepository.save(user);
-        User user2 = new User("NewUser2", "mail2@email.com", "89101112", "5678","5678");
-        userRepository.save(user2);
-        User user3 = new User("NewUser3", "mail3@email.com", "13141516", "9101", "9101");
-        userRepository.save(user3);
-        return "ok";
-    }
-
     @GetMapping("")
     public Iterable<User> getUsers (){
         return  userRepository.findAll();
