+++ src/main/java/org/launchcode/inventorytrackergradle/models/Item.java
@@ -5,6 +5,9 @@ import com.fasterxml.jackson.annotation.JsonBackReference;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 @Entity
 public class Item extends AbstractEntity{
@@ -14,6 +17,14 @@ public class Item extends AbstractEntity{
     @JoinColumn(name = "manufacturer_id")
     private Manufacturer manufacturer;
 
+    @NotNull
+    @NotBlank
+    @Size(max=75)
+    private String name;
+
+    @NotNull
+    @NotBlank
+    @Size(max=250)
     private String description;
 
     private int numberInInventory;
@@ -23,9 +34,10 @@ public class Item extends AbstractEntity{
     public Item() {
     }
 
-    public Item(Manufacturer manufacturer, String description, int numberInInventory, int numberMinimumToKeepOnHand){
+    public Item(Manufacturer manufacturer, String name, String description, int numberInInventory, int numberMinimumToKeepOnHand){
         super();
         this.manufacturer = manufacturer;
+        this.name = name;
         this.description = description;
         this.numberInInventory = numberInInventory;
         this.numberMinimumToKeepOnHand = numberMinimumToKeepOnHand;
@@ -38,6 +50,14 @@ public class Item extends AbstractEntity{
 
     public void setManufacturer(Manufacturer manufacturer) {
         this.manufacturer = manufacturer;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
     }
 
     public String getDescription() {
